Update some action buttons to disambiguate 'Cancel' (#243)
* Update some action buttons to disambiguate 'Cancel'

This change replaces red 'Cancel' actions with violet 'Back' actions on displays where it is intended as 'Cancel this action' rather than 'Cancel this booking'.

* Add missing colour Violet to enumerations

* Reduce some duplication of props in CtaButton and ConfirmModal

This changes the `CtaButton` to export its props, then references those in the `ConfirmModal` props. This change allows complex restrictions like cta colour to be passed through without having a duplicate/mismatching type at a higher level. A downside of this is that it uses an import style that mismatches the component import from `index.js`. I think this is worthwhile, but I'm open to challenge.

* Update admin cancel booking CTAs to match user cancel

This changes the text and colour of the CTAs for an admin deleting a booking to match those when a user does.